feat: finalise photo stream
