+++ .eleventy.js
@@ -5,52 +5,7 @@ const pluginSyntaxHighlight = require("@11ty/eleventy-plugin-syntaxhighlight");
 const pluginNavigation = require("@11ty/eleventy-navigation");
 const markdownIt = require("markdown-it");
 const markdownItAnchor = require("markdown-it-anchor");
-const ExifReader = require('exifreader');
-
-const PHOTOS_DIR = 'img/photos'
-const photos = fs.readdirSync(PHOTOS_DIR);
-
-const photoCollection = photos.map((filename) => {
-  return {
-    filename,
-    name: filename.split('.jpeg')[0],
-    src: `/${PHOTOS_DIR}/${filename}`,
-    path: `${PHOTOS_DIR}/${filename}`,
-  }
-})
-
-/* Parse date string in YYYY-MM-DD hh:mm:ss format
-** separator can be any non-digit character
-** e.g. 2017:03:09 14:49:21
-*/
-function parseExifDate(s) {
-  var b = s.split(/\D/);
-  return new Date(b[0],b[1]-1,b[2],b[3],b[4],b[5]);
-}
-
-async function readPhotosWithExif() {
-  const photosWithExif = []
-  await Promise.all(photoCollection.map(async (photo) => {
-    const exifData = await ExifReader.load(photo.path)
-    const lensFocalLength= (exifData.FocalLength?.value.toString() || '').split(',1')[0]
-    const date = exifData.DateTimeOriginal?.description ? parseExifDate(exifData.DateTimeOriginal?.description) : parseExifDate(exifData.DateTime.description)
-    photosWithExif.push({
-      ...photo,
-      date,
-      make: exifData.Make?.description || '',
-      model: exifData.Model?.description || '',
-      lens: exifData.Lens?.value || '',
-      lensFocalLength: lensFocalLength || '',
-      lensFocalLengthEquivalent: (lensFocalLength && lensFocalLength  * 1.5) || '',
-      iso: exifData.ISOSpeedRatings?.value || '',
-      aperture: exifData.FNumber?.description || '',
-      // 
-    })
-    // console.log(exifData)
-  }));
-  // console.log(photosWithExif)
-  return photosWithExif
-}
+const readPhotosWithExif = require('./photoReader')
 
 module.exports = function(eleventyConfig) {
   // Add plugins
@@ -67,11 +22,11 @@ module.exports = function(eleventyConfig) {
   eleventyConfig.addFilter("readableDate", dateObj => {
     return DateTime.fromJSDate(dateObj, {zone: 'utc'}).toFormat("dd LLLL yyyy");
   });
-  
+
   eleventyConfig.addFilter("monthYearDate", dateObj => {
     return DateTime.fromJSDate(dateObj, {zone: 'utc'}).toFormat("LLLL yyyy");
   });
-  
+
   eleventyConfig.addFilter("aperture", fNumber => {
     return fNumber === 'f/0' ? '' : fNumber
   });
@@ -102,18 +57,16 @@ module.exports = function(eleventyConfig) {
   // Copy the `img` and `css` folders to the output
   eleventyConfig.addPassthroughCopy("img");
   eleventyConfig.addPassthroughCopy("css");
-  
+
   eleventyConfig.addPassthroughCopy("CNAME");
 
   eleventyConfig.addPassthroughCopy({
-    "node_modules/github-markdown-css/github-markdown.css": "css/github-markdown.css" 
+    "node_modules/github-markdown-css/github-markdown.css": "css/github-markdown.css"
   });
 
   // Create collection of photos
   // TODO: assess processing these images for web https://www.raymondcamden.com/2021/04/07/building-a-simple-image-gallery-with-eleventy
-  eleventyConfig.addCollection('photos', async function(collectionApi) {
-    return await (await readPhotosWithExif()).sort((itemA, itemB) => itemB.date - itemA.date);
-  });
+  eleventyConfig.addCollection('photos', readPhotosWithExif);
 
   // Customize Markdown library and settings:
   let markdownLibrary = markdownIt({
