+++ css/index.css
@@ -50,7 +50,18 @@ figcaption {
 }
 
 figure.photo {
-  margin: 8rem 0;
+  margin: 2rem 0;
+}
+
+@media (min-width: 767px) {
+  figure.photo {
+    margin: 8rem 0;
+  }
+
+  figure.photo:first-of-type {
+    margin-top: 2rem;
+    margin-bottom: 8rem;
+  }
 }
 
 /* figure.photo .photo-image {
